requirements.txt
streamlit>=1.28.0
pandas>=1.5.0
plotly>=5.15.0
numpy>=1.24.0
openai>=1.0.0
python-dotenv>=1.0.0
requests>=2.31.0
reportlab>=4.0.0
matplotlib>=3.7.0
markdown2>=2.4.0
weasyprint>=60.0